Use mixin and method injection for latex parser
